Adapt for pseudoconstant changes in CiviCRM 4.4
Fixes issue #40: "'9000' is not a valid option for field
payment_instrument_id" when saving after editing an existing SDD PP
record.

(Doesn't affect saving on initial creation, which is still broken
because of #37...)

Note: this commit fixes the extension for 4.4; but it won't work for 4.3
anymore... I can't see a (sane) way to make it work with both.

The noise in the generated DAOs is from changes in the 4.4 GenCode
script -- the only functional change is in SEPACreditor.php , reflecting
the schema change in Creditor.xml .