One SDD PP type is enough
It's nice to have a PP type for SDD created when the extension is
installed. It's not so nice though to have two of them created, with
slightly different parameters...

By the right of the eldest, I decided to keep the original one, and drop
the upstart. (Hi CRM/Core/Payment/SEPA.mgd.php )