Fix retrieval of `creditor_id` in batchContribution()
Don't try to find the `creditor_id` from the `payment_instrument_id`,
which is no longer related in the New World Order.

Conveniently, the `creditor_id` is actually available directly from the
mandate object, which we get as a parameter... No need for any extra
queries.

This also drops the gross hack of creating a fake creditor on the fly.
No idea what that was for, but it doesn't apply anymore, and I *very*
much hope it's not needed. (Works fine here without.)