Base fallback Mandate Reference on Mandate ID rather than Entity ID
The previous code had several issues: the conditional was buggy; the
generated reference was actually not guaranteed to be unique (as it
didn't include the Entity Type); and the random case was not random
enough.

While all of these issues could have been fixed, basing the reference on
the Mandate's own unique ID is more generic (no conditionals); simpler;
probably safer; more obvious (especially once a mandate can be used for
multiple contracts...); and IMHO also more informative.

As the auto ID is only available once the record was created, the code
needed some shuffling around: we initially create the record with an
arbitrary unique reference (to satisfy the index constraint); and only
later update it with the proper meaningful reference.