SepaMandate::get() API: Drop "entity_table" default value
Drop the default, and update all callers that relied on it.

I can't see why it would be useful to have such an arbitrary default
here. All it does is breaking things in ways that would be painful to
work around. (For example when doing lookups by `reference` rather than
`entity_table`+`entity_id`.)

This fixes the mandate links on the dashboard for OOFF contributions --
and probably other things too...