Display mandate data in contribution view for OOFF contributions
For OOFF contributions, we don't have a recur record -- so we can't
display the mandate data on the ContributionRecur view, but rather have
to put in on ContributionView.

The code paths used here are pretty much separate from the recur ones,
which does cause some redundancy, but it is for the best: the one-off
case uses a significantly different template, as well as a different
approach for injecting it in the output page -- both of which are major
improvements. (Technically as well as in the user-visible result.) Going
forward, the recur case probably should be aligned with the approach
used here.