Make Transaction Group Reference unique as well
This was less problematic than the File Reference, as it includes the
unique Creditor ID, and the collection type (both of which wouldn't work
at file level); however, it still suffered from the possibility of
having several groups with the same due date. Thus, we need to include
the auto ID here as well.