Leave default <RmtInf> empty
In want of code to generate a useful generic default, rather than
setting an arbitrary constant string, don't set anything at all for now.

Also, if nothing is provided, suppress creation of the element in the
XML file alltogether. Don't know whether this is really necessary -- but
it's probably safer that way; and it saves some ballast in the file
too...