Dasboard: Create separate table for each Creditor
Instead of displaying one large table with all Transaction Groups of all
Creditors, sort the Groups by Creditors, and display a separate table
(in a collapsable "accordion" tab) for each Creditor.

This is probably a useful change in general; and it will be crucial for
on-demand batching.

(Note: The grouping should actually be done by Creditor Tag rather than
Creditor -- but this seems a rather obscure feature; so punting it for
now...)

To enable the sorting, the SepaTransactionGroup::getdetail() API call is
extended to include the Creditor ID in the result.