XML generation: Store EndToEndId (`trxn_id`) in database
Permanently record the ID we put in the file, rather than only
generating it on the fly. (Fixes issue #106.)

This way we always know the actual ID that has been submitted to the
bank, so we can match it easily and reliably. (Generating on-the-fly is
not reliable, as that might change when changing the Creditor Prefix for
example.)

Also, it's visible in the CiviCRM UI this way; and it can be used in
searches.