WIP: Always use BAO to update contribution status
The only case we used where the API update actually works, is setting to
'Completed'. That one is problematic in other ways though, as it needs
setting the date, which requires a rather tricky approach (using an
extra invocation of SepaTransactionGroup.get). This is rather obscure;
quite inefficient; and what's more, it won't even work as soon as we
introduce another nesting layer before the Contribution.create -- which
we will need to implement filtering by previous Contribution status...

All in all, unless the Contribution API changes to actually accept all
the needed status changes, the uniform BAO approach is indeed cleaner
and simpler.