WIP: Don't update status of contributions that have an unexpected status
To make the actual check work, we need to make the 'from_status'
parameter mandatory in the updatestatus API, and pass it to the backend
function explicitly.

Open question: is it actually a good idea to silently ignore an
unexpected status? This will create an inconsistent SEPA history -- when
SEPA contributions are modified behind the extension's back, bad things
will happen no matter what. Perhaps we should error out instead of
ignoring -- or just keep the previous behaviour of forcefully updating
the status regardless...

ToDo:
- more comments?