DashBoard: Make the entire contribution group rows clickable
Instead of limiting the clickable area for activating the detail listing
to a few select fields, make the entire table row clickable.

This is more discoverable, more convenient, and even sligtly simpler in
the implementation. The only minor downside is some possible confusion
about the file links contained within the rows... Still, this should be
acceptable in view of the major benefits.