WIP: Fixup: Drop bogus conditional when setting default `sequence_number`
This never could have worked, because of a mistyped variable name.
(Whoops...)

As it turns out, the conditional was pointless anyways: the API code
setting the custom fields is executed *after* the main entity's post_
hook fires -- so we actually never have an existing value here. Rather,
we can just set a default value unconditionally; and it will be later
overwritten by any value explicitly requested through the API.

- To be folded into original generate recur / sequence_number commit