WIP: Drop batching hook dispatcher
With the new batching approach, we do not rely on hooks anymore. Make
this explicit by removing the machinery for invoking such hooks.

Note: the equivalent mandate hook machinery is still in use and left in
place for now -- though we might want to replace the automatic magic
with a hard-wired dispatcher at some point, to make it easier to
follow what is actually happening when...