WIP: Kill broken Membership Payment forms
While we tried not to break this functionality gratuitously, we also
didn't spend any serious effort on keeping it working with the new
batching approach etc. -- and unsurprisingly, it's not.

Remove it for now to avoid user confusion. Might be fixed and readded
later, if it's still deemed useful by anyone. (Todo: check whether this
is covered by the existing upstream membership payment functionality?)