WIP: PP: doDirectPayment(): Get Creditor from passed `payment_processor_id`
There is no need to rely on passing the Creditor through a global
context from the validateForm hook, as the Payment Processor used is
always available as an explicit parameter in doDirectPayment(), which is
what we get the Creditor from anyways.

Unfortunately, we still need to pass the `payment_instrument_id` in the
context, as there is no other way to know that we are dealing with a
SEPA contribution in the Contribution create pre hook.