WIP: Allow entering new SDD Contributions through back-office interface
The existing "Add Credit Card Contribution" back-office form is
essentially the canonical method for adding new PP-based Contributions.
It only supports Credit Card Providers out of the box, because that is
what the PP mechanism itself was primarily built for; but it turns out
it can be made to work with our SEPA Processor too, using some form hook
hackery (not more than for the Online Contribution Pages) -- thus
providing a streamlined UI handling SDD just like other types of
automated payments, with fairly little new code.

We expect that this approach will enable SDD-based Membership Paymennts
as well without too much extra work.

Note: the form currently filters out any PPs not declared as Credit Card
Providers. To make it admit the SEPA PP as well, we needed to change the
`payment_type` of the SEPA Payment Provider to "Credit Card". While this
is a misnomer and probably quite confusing, it doesn't appear to cause
any actual ill effects...