Allow setting execution date for one-off Contributions
While the back-office form for entering new PP-based contributions has a
Start Date field for Recurring Contributions, it doesn't have an
equivalent field for setting an explicit Receive Date for single
Contributions. However, the Start Date field is actually also available
when not selecting the recurring option; and it seems an obvious choice
to use it as the execution date for single Contributions as well.

The value indeed already is being passed by CiviCRM as `receive_date` to
the PP callback -- but not to the Contribution creation. Propagating it
explicitly from the PP appears to be sufficient to make this work fully.

This is a PP-specific workaround of course. Most likely the incomplete
support should be considered a bug ideally to be fixed in CiviCRM
itself...