WIP/Hack: Optionally create separate files for each group
Some banking systems fail to process SEPA XML files with more than one
group. (Although these are perfecty valid... Hurray for "standards".)

To deal with such broken systems, it is now possible to choose whether
to put all upcoming groups in one XML file, or generate a separate file
for each.

The actual setting is hardcoded for now. (For DiCo, we keep the much
more convenient variant with all groups in one file...)