WIP/Fixup: Define default value for `sequence_number` field
When using the back-office form to enter new contributions, the value
provided for `sequence_number` in the input field will always overwrite
the default value we set in hook_post_contribution_create(), because the
custom value is stored only after the main entity is stored (and the
main entity's post hook invoked). If we don't have a value set in the
form, a NULL value will be saved.

Giving the `sequence_number` field a default value of '1' makes sure the
right value is stored (unless the user filling the form meddles with
it).

Note: the custom field definition change only takes effect during
installation. For existing deployments, the custom field needs to be
updated accordingly by hand.

- to be folded into commit introducing the `sequence_number` field