Normalise IBAN and BIC before storing in DB
This avoids generation of invalid SEPA XML files when certain types of
unnormalised values are entered, e.g. using lowercase letters. (There
was already some normalisation during the actual file generation -- but
it was incomplete; and this is really the wrong place for doing the
normalisation...) Also the values in the DB are less chaotic this way.

The normalisation is done in the BAO layer, so it will work no matter
how the data is entered. (frontend/backend/import)