WIP: Work around different form processing in CiviCRM 4.5
CiviCRM 4.5 introduced the new "Chain Select" mechanism, which turns out
not to be robust against form fields being removed.

To get around this, when swapping out the CreditCard payment fields for
the SEPA DD ones during submit processing, instead of actually removing
the original fields, just make them non-mandatory. While this is
somewhat less elegant IMHO, the original fields are not really in the
way when not mandatory...