Check mandatory BIC for non-domestic payments
When checking IBAN/BIC validity, if no BIC was supplied, check whether
IBAN-only is actually allowed for the given IBAN, i.e. whether we are
dealing with a domestic payment.

The check is performed by comparing the country code in the IBAN to the
one in the creditor IBAN. This means we have to retrieve the Creditor
IBAN first, which required some code restructuring.