Perform IBAN/BIC validation on Creditor data too
Add the validation/normalisation calls in the Creditor BAO, as well as
the PP/Creditor edit form.

As the IBAN-only check makes no sense in this case (and there is no IBAN
to check against), this required making the check optional at the code
level.

Note that the validation error is not displayed correctly in the form:
the form bounces, but no error message is shown. I have no idea what
causes this -- I suspect something fishy in the upstream PP edit form...