WIP: Add batching option to generate separate files for each Sequence Type
For completeness, add an option to group only different dates but not
different Sequence Types (or Instruments), to complement the existing
options to not group anything; to group everything using the same
instrument; or to group everthing into a single file.

There might or might not be a use case for this -- but at any rate, it
makes the code more regular, and thus probably easier to follow.

- To be folded into original commit introducing batching modes? (And the
  new batching mode introduced with COR1 handling.)