WIP: Adapt Underscore/Lo-Dash inclusion for CiviCRM 4.5 ways
In CiviCRM 4.5, underscore.js was replaced by lodash.js. Also, lodash.js
is now always included in every page served by CiviCRM -- and if we
still try to load it explicitly with addScriptFile(), it leads to a
situation where the system include is automatically omitted to avoid
duplication, which in turn causes failures because the explicitly
requested one is loaded too late for other system scripts...

In other words, if we want to use the CiviCRM-provided copy of Lo-Dash,
we *have* to use the system instance, rather than trying to load it
ourselves. (Plus, it's more efficient anyways.)

Note that this change breaks CiviCRM 4.4 compatibility. I can't think of
a way to make it compatible with both without conditionals or some other
form of non-trivial trickery...