Fix a "strict warning"
Apparently having an assignment in the argument list of a function call
for a parameter passed by reference produces a "strict warning: only
variables should be passed by reference". I'd consider that a bug in
PHP, but whatever...

Now we just pass an uninitialised dummy variable for the unused
parameter instead of explicitly initialising it to null -- IMHO that's
less clear, but it works and avoids the warning.