Revert "WIP: Work around different form processing in CiviCRM 4.5"
This reverts commit 8725341c30d2e4d00a2c6f848c206de6e5f34b0a.

Turns out keeping the billing fields (and only making them
non-mandatory) has a somewhat nasty side effect after all: whenever
creating a PP Contribution through the back-office form, the Contact
grows a "Billing Address". (With values copied from the main Address
and/or the Contact's name.)

As the original problem with chain-selects that prompted this change has
been fixed upstream in newer versions of CiviCRM 4.5.x, we can just
revert to the original approach. This is not exactly clean either (it
produces some PHP notices), but at least we are not aware of any nasty
side effects with that... (Let's hope Eileen's changes in CiviCRM 4.6
will enable us to get rid of all these nasty hacks.)