Respect `contribution_recur.installments`
If a recurring contribution has been created with a limited number of
installments, make sure we do not create any new installments past that
number.

Note: For now, we just avoid creating extra contributions; but we do not
change the status of the recurring contribution to 'completed'. This
will need to be handled along with other recurring contribution status
handling, which is missing alltogether so far.

(Open question: should the status change already when the last
installment is created, or only when it is marked 'completed'? The
latter is more complicated to implement, but probably more useful
too...)