Fix Payment Instrument in accounting records with CiviCRM 4.5
With CiviCRM 4.5, we need to explicitly pass the `payment_instrument_id`
back in when updating the contribution status to 'completed', as
otherwise the `financial_trxn` records are created with a NULL
`payment_instrument_id`.

Also provide upgrade code to automatically fix up the wrong entries
generated by previous versions out in the wild.