Introduce wrapper class to fix transaction rollback on exception
Add a wrapper around CRM_Core_Error, which does a rollback rather than a
commit when the destructor is invoked. (Most notably when exceptions are
thrown.)

All uses of CRM_Core_Error are updated to use the wrapper instead.

This works around the strange/buggy CiviCRM upstream behaviour. In
CiviCRM 4.6, the API grows an optional new way of managing transactions,
which apparently avoids this issue -- so we probably can get rid of this
workaround once we switch to 4.6...