Drop '+' modifier from date_create_from_format() format string
It turns out this is not available in old versions of PHP (such as the
one in use at KDE) -- which is *not* mentioned in the PHP documentation
:-(

At least with the current code, the '+' doesn't actually seem to be
necessary though, as all existing callers should pass pure dates in the
correct format. Making it explicitly reject wrong formats (such as
datetime combinations) is probably a good thing...

(Another option would be using date_create() instead, i.e. no explicit
format. This appears to work fine as well in my testing... Any time or
timezone passed in does not actually affect the outcome, as we only do
maths on the date part, and only return that at the end.)