<MsgId>, <PmtInfId>: Generate according to new scheme, avoiding overflows
Since we started putting the SEPA Instrument (CORE/COR1), the Sequence
Type (FRST/RCUR/OOFF) and the Due Date in the <MsgId>, and also the
Instrument in the <PmtInfId>, the verbose IDs generated up till now can
get too long, and overflow the allowd field length of 35 characters.

As the fields actually present in the <MsgId> depend on the batching
mode, we didn't experience this problem at Digitalcourage, where we only
batch the Instruments seperately, while all Sequence Types and Due Dates
are combined in one file, and thus do not show up in the <MsgId>.
However, when batching each Group into a separate file, the <MsgId> gets
way too long.

The <PmtInfId> always contains the same fields; but it's generally
shorter (as it only carries one date), and thus (barely) fit within the
bounds thus far -- but would cause problems at Digitalcourage very soon,
when the group IDs reach 1000, and thus the field becomes wider...

To alleviate these problems, we are switching to a different scheme for
generating the IDs, employing various abbreviations. This gives us ample
reserve for the automatic IDs getting wider; as well as some extra room
for somewhat longer Creditor Tags / Mandate Prefixes.

The SDD XML file names so far were directly derived from the <MsgId>.
However, as the file names are not limited in length, we do not need to
apply the abbreviations here, and can keep the old scheme instead. This
means though that rather than deriving them from the generated <MsgId>,
we now need to generate them separately.