SDD XML file names: Use same separator as in <MsgId>
Thus far, the file names used '_' as separator, while the internal IDs
used '-'. ('_' is not a valid character in SEPA files.)

These is no real rationale for this difference though -- it probably
just causes extra confusion...

Thus, changing the separator in the file names to match the one in the
internal IDs.