Don't limit `mandate_prefix` to 4 chars
This was a totally arbitrary and way too low limit.

It's hard to say what a reasonable limit is though, as it depends on the
exact format of the various references it is used in, as well as the
number of digits in the automatic IDs. Thus, we just use the maximum
size of the references in the XML files (35 chars) where it appears. Of
course the prefix alone can't fill the entire IDs -- but setting any
other fixed limit would be quite arbitrary...

Also set the 'maxlength' for the form field on the Creditor (PP) edit
form. (But set a much smaller 'size', to give a hint that this should
actually be a fairly short prefix...)

While at it, also improve the title and comment of the DB field.