Fix: Avoid creation of surplus "SEPA File Formats" Option Values
Using chained API calls for creating Option Values along with the Option
Group as a "managed" entity doesn't work, as the chained calls are
invoked incorrectly when the managed entities are disabled/re-enabled.
(And also when they are removed, once we get to this...)

To fix this, we now create the Option Values as separate entities
instead.

To allow for this, we need to figue out the Option Group ID before we
build the "managed" entries in the _civicrm_managed() hook. The only way
we can do this is by creating a placeholder Option Group in advance.
This is not very elegant, but seems to work just fine.

Also added an upgrader action to purge any surplus Option Value entries
already present in the DB, and switch the correct ones to "managed"
entities.