Turn the 'Batched' Option Value (Contribution Status) into a "managed" entity
This way, it doesn't use the ugly and buggy custom option installer; and
its presence and `is_active` status are automatically managed when
disabling, re-enabling, or deinstalling the extension.

`cleanup` is set to 'unused', i.e. the entity will be automatically
removed during uninstallation of the extension, unless there are
Contributions using it.

As a side effect, the Option Value is marked `reserved` now, as it ought
to be.

This commit includes upgrader actions for both changes as well.