Handle `sequence_number` on extension disable/enable/uninstall
We have to manage the state manually through the relevant hooks, because
Custom Groups do not work as "managed" entities (due to incompatible API
behaviour); but also, because the behaviour we are after is actually
different from the one implemented for "managed" entities.

The Group is only deactivated when there are no Contributions having the
Custom Value set. Otherwise, we want to keep the existing "Sequence
Number" values visible, even when the extension is disabled (and
possibly uninstalled).

Likewise, the Group is only removed on extension uninstall if there are
no users.

With the preceeding commits, `sequence_number` should only be in use if
it is actually needed; so it won't be sticking around if not necessary.