Generate upcoming: Catch missing Sequence Number
Safety measure for the `sequence_number` default handling changes: check
whether a valid Sequence Number is present, and bail out if not.

Unfortunately, the Exception is not very visible when running in a cron
job... Still, hopefully it will be noticed sooner or later.

Note that this safety check might also trigger in some other error
situations.