sepa.sql: Apply 'IF NOT EXISTS' to all tables
While it was already present for some tables, others were missing it,
preventing re-installation of the extension.

Another approach would be to explicitly check whether the tables are
already present from PHP code, and skip the execution of the entire file
if that's the case. That however would prevent use of the auto-install
feature from the civix "upgrader".

OTOH, adding the extra SQL clause makes importing the generated SQL more
complicated... The jury is still out on this.