Automatically upgrade DB in sepa_civicrm_managed() hook
This prevents the DB from being messed up if the hook is invoked by
something before the upgrade has been triggered manually.

This will also be used when adding support for upgrading from older
versions of the extension after the upcoming rename.

To allow for the new entry point, the actual upgrade code is factored
out into a dedicated class invoked both by the new automatic backgroud
upgrade, and the old extra page.

The extra page is not strictly needed anymore, as the automatic upgrade
can be forced by invoking civicrm/menu/rebuild. However, the extra page
is more convenient for debugging -- so leaving it in for now.