Provide automatic upgrades from 'org.project60.sepa'
Add some extra handling to allow installing the "new" extension
('sfe.ssepa') on top of an existing installation of the old one
('org.project60.sepa').

The old extension is automatically marked as uninstalled, while the
existing data is taken over and upgraded.

(Before installing the new extension, the old one needs to be disabled
by hand, but *not* uninstalled -- the hijacking relies on the old
extension still being present.)