Implement a new upgrade mechanism
New approach based on the standard CiviCRM upgrade mechanism.

We still allow invoking it explicitly through civicrm/sepa/upgrade
though, because there doesn't seem to be another way to trigger DB
upgrades from the UI, if the code is upgraded on the FS level rather
than through the UI. (Though it can be done through the API explorer...)

Keeping the old upgrade mechanism in place for now -- though it's not
fully tested for interaction with the new approach (especially regarding
the rename handling), so not sure whether it's really still possible to
upgrade directly from versions < 0.3...

Note: While this new approach is based on the standard CiviCRM upgrade
mechanism, it doesn't use the auto-generated code from civix, which is
based on linear numeric version numbers. Rather, it checks the actual DB
state for each upgrade task, just like with our old mechanism.

This commit doesn't add any actual upgrade tasks yet -- it just puts the
framework in place, to be used for upcoming upgrades in Ssepa 0.4 and
beyond.