doDirectPayment(): Return $params
While $params is passed in by reference, and thus callers can in theory
get any modified values from there, doDirectPayment() is actually
expected to pass it back as the function return value as well.

With CiviCRM 4.6, some code paths rely on the return value even for
basic contribution parameters that are not modified by
doDirectPayment(), and will fail if we don't pass them back.