sepa_civicrm_entityTypes(): Fix typo in class name
The `SepaTransactionGroup` entry was accidentally(?) listing the BAO
instead of DAO class. While I still don't know what these "entity types"
are actually used for, this suddenly started breaking some of the magic
API calls on this entity starting with one of the CiviCRM 4.6.x stable
releases, which in turn broke our XML generation -- see
https://github.com/sfe-ev/ssepa/issues/4

Switching this to the proper DAO class seems to fix the problem.