GP-14437 Fix default SEPA payment instrument not being set
This fixes an issue where the default SEPA payment instrument is not
offered as an option when creating a new creditor if there is no
existing creditor using the FRST/RCUR payment instrument. This was
due to CRM_Sepa_Logic_PaymentInstruments::getAllSddPaymentInstruments
only loading available payment instruments that are already used by
a creditor.